requirements.txt
qrcode
cryptography
mysql-connector-python
opencv-python
pyzbar